Refactor read_files to check for Markdown
Related to a bug where we tried to read in a `.DS_Store` file, I added a
check to for the file extension and verify that it's a `.md` before
attempting to `slurp` in the file.

That check started failing on actual Markdown files so I added a better
`is_markdown` function to check the extension via `Path`.