Refactor: Add subcommands
Adds the basics for supporting subcommands:
- compile
- word-count
- obsidian

The goal of this is to expand a little bit of the functionality,
primarily to automate the installation (or removal?) of the integration
with Obsidian. In theory, it establishes a pattern for adding
integration support to other editors, though I don't have any planned.