Refactor compile with fewer options + multi-manuscript generation
I made the executive decision to simplify the CLI. Rather than
specifying anon, font name/size, etc on the command line, running
compile will now generate several manuscripts in the output directory,
in Courier New and Times New Roman (12pt font), in Classic and Modern
Standard Manuscript Format.

This is, in part, to limit the number of shell commands we need to
setup, at the expensive of a little flexibility.

The ultimate goal of this is to cover the 80% of use-cases, which is
following one of the two Shunn formats, with one of the two preferred
fonts.

If someone wants a MS in a different format, it'll be up to the writer
to open the MS, make the changes, and save it to a new file name.